// src/common/enums/cutoff-type.enum.ts
export enum CutoffType {
    BI_WEEKLY = 'BI_WEEKLY', // Cutoff period of two weeks
    MONTHLY = 'MONTHLY', // Cutoff period of one month
    WEEKLY = 'WEEKLY', // Cutoff period of one week
    DAILY = 'DAILY', // Cutoff period of one day
    ANNUAL = 'ANNUAL', // Cutoff period of one year
    QUARTERLY = 'QUARTERLY', // Cutoff period of three months
    SEMI_ANNUAL = 'SEMI_ANNUAL', // Cutoff period of six months
    BI_ANNUAL = 'BI_ANNUAL', // Cutoff period of two years
}